[bugfix] view::iota_minus overflow for UnsignedIntegral
The problem:

iota_minus was implemented as (D)(v0 - v1), where D is the
iota_difference_t of the UnsignedIntegral type of v0 and v1.

Example: v0 = 0, and v1 = 1 causes iota_minus to overflow.

Note: this overflow is perfectly defined behavior, but the result is
not what the users of iota_minus expect.

The fix:

v0 < v1? - static_cast<D>(v1 - v0) : static_cast<D>(v0 - v1);

That is, we check which integer is larger, we perform a non-overflowing
substraction using the current type, and then we cast to the difference
type and correct the sign.

Pros: it maximizes the range for which iota_minus works correctly
without overflowing.
Cons: it introduces a branch and a signed bit change.

There is an alternative solution that works in clang and gcc but
was not pursued:

static_cast<D>(v0) - static_cast<D>(v1)

That is, we truncate the UnsignedIntegers into the different type and
then perform the substraction.

Pros: simple and fast.
Cons: AFAIK this truncation is implementation defined.

Given:

v0 = std::numeric_limits<UnsignedIntegral>::max()
v1 = std::numeric_limits<UnsignedIntegral>::max() - 1

clang and gcc implement the truncation as

static_cast<D>(v0) == -1
static_cast<D>(v1) == -2

and thus (-1) - (-2) == 1 which is the correct result.

This solution is simpler, and works reliably in all compilers that
range-v3 supports for now. Maybe it should be pursued in the future.