rm virtual keyword in override
what good is it stating an override is virtual? the base class,
std::bad_cast, already says the what() member function is virtual.

if the base class' signature is virtual, then the keyword is a no-op. if
the base class' signature were not virtual, then the virtual override
can bring confusion down the line.