Specialise Boost Range metafns for range-v3 views (#495)
* Specialise Boost Range metafns for range-v3 views

This allows range-v3 views to be treated as ranges by the Boost Range
library and other related Boost libraries.

This is an attempt to tackle ericniebler/range-v3#486.

* Name BOOST_RANGE_HOOK > RANGES_SATISFY_BOOST_RANGE

Rename BOOST_RANGE_HOOK() macro to RANGES_SATISFY_BOOST_RANGE() as
suggested by ericniebler.

* Trivial tweak to trigger Travis-CI build for P-R

The pull-request's build failed but just due to a time-out. This trivial
commit should trigger a new build.