Port subrange from cmcstl2
Fixes #766

Feature detection updates:
* Detect `RANGES_CXX_IF_CONSTEXPR`
* Detect `RANGES_CXX_ATTRIBUTE_NODISCARD`. `RANGES_NODISCARD` is defined as `[[nodiscard]]` when supported, and empty otherwise.
* New diagnostic control `RANGES_DIAGNOSTIC_IGNORE_CXX17_EXTENSIONS` to suppress the warning clang emits when you use `[[nodiscard]]` in C++11 mode *despite that it defines `__has_cpp_attribute(nodiscard)`*.

Better realize a notion I've mentioned before that `subrange`/`ref_view` are special in that they borrow the iterators of the underlying range, so the iterators can safely outlive the view. These views are not constructible from rvalue containers, so they participate in the range lifetime rules exactly like lvalue references. A user who passes a reference to an lvalue container to the Range library must guarantee that the container's lifetime encompasses the lifetime of the reference which may be stored. The library assumes that rvalues could be temporary and never stores references to them beyond the duration of a function call.

The `ViewableRange` concept encapsulates this notion for range arguments: a range is "viewable" if it is already a `View` or if it is an lvalue. Only `Viewable` ranges have lifetime guarantees strong enough to participate in the view composition algebra. The new poorly named `ReferenceableRange` concept encapsulates the opposite notion for return values: only `Referenceable` ranges have lifetime guarantees strong enough to ensure that a returned iterator does not dangle. For now, that means lvalues, `subrange`s, and `ref_view`s. (I've not admitted `iterator_range` because that type becomes deprecated with the introduction of `subrange`, and used in range-the tests specifically as a non-`Referenceable` range.) The set of `Referenceable` ranges is opt-in via the user-specializable `is_referenceable_range` trait.

Altogether, this means that iterator-returning algorithm overloads that take range parameters `R` should return `meta::if_<ReferenceableRange<R>, iterator_t<R>, dangling<iterator_t<R>>>` (`safe_iterator_t`) and range-returning algorithm overloads that take range parameters should return `meta::if_<ReferenceableRange<R>, subrange<iterator_t<R>>, dangling<subrange<iterator_t<R>>>` (`safe_subrange_t`).

Cleanup `all_fn` a bit:
* Always test `uncvref_t<T>` for `View`-ness instead of `T`
* Properly return `uncvref_t<T>` from the reference-wrapper-to-`View` unwrappers
* Specialize `is_referenceable_range` for `ref_view`

Steal the name `PairLike` from `map.hpp` for `subrange` (rename to `HasFirstAndSecond`)

Some comments on the port:
* Yes, `PairLike` really needs to be split into that many pieces. The new Standard-mandated `static_assert` in `get` makes writing generic code around `pair` and `tuple` very "interesting".
* Yes, the constraints on the `subrange` class template itself seem to be necessary. Without them clang is confused by the deduction guide tests: it's not clear to me if this is a Clang bug or confusion on my part around deduction guides.
* The fact that `range_size_type_t` is unsigned even when `ranges::size` returns a signed type is becoming more of a problem - I had to insert some creative `static_cast`s in various places to suppress warnings after the `iterator_range` => `subrange` replacement.
* This implementation doesn't need separate sets of "Lvalue range" and `subrange` converting constructors: it has a single set of converting constructors that accept `ReferenceableRange`.
* The second parameter of `sized_subrange` defaults to the first parameter. This seems convenient.
* As I mentioned above, `safe_subrange_t<R>` in this implementation returns either `subrange<iterator_t<R>>` or `dangling<subrange<iterator_t<R>>>`, whereas cmcstl2 uses `tagged_pair<tag::begin(dangling<iterator_t<R>>), tag::end(dangling<iterator_t<R>>)>` in the dangling case; I prefer to preserve the fact that the value that dangles *is* a range.

Note that I've replaced uses of `iterator_range` with `subrange` in the headers, but tried to touch the tests as little as possible. This demonstrates that the conversions from `subrange` and `ref_view` work well at minimizing potential user breakage due to the replacement. The only likely source of fallout is - as in the tests - users passing `view::all(lvaluerange)` and calling member `get_unsafe` on the return value.