Added explicit options for enable or disable testing
These options are visible from the cmake-gui and cmake-curses-gui;
their presence is more discoverable than the local variables used
previously.

In addition, the default values for these options is contingent
upon whether range-v3 is the highest level project. When range-v3
is composed in another project (as a git submodule, for instance)
tests are not built and no option appears in the GUI.