Flip CPP_true to CPP_false to avoid IFNDR
`enable_if_t<[non-dependent] && [dependent]>` is IFNDR when the non-dependent subexpression is `false`, regardless of the value of the dependent expression. `enable_if_t<[non-dependent] || [dependent]>` is well-formed regardless of the value of the non-dependent expression, and equivalent to the prior form if the dependent expression is always false rather than always true.