Work around broken integration with latest libc++. (#1635)
* Work around broken integration with latest libc++.

In newer versions of libc++, the base template of std::iterator_traits
provides a member typedef called __primary_template which is an alias
to the std::iterator_traits specialization itself. This fix works with
both the old version of libc++ and the new one.

Fixes issue #1633.

* Fix is_std_iterator_traits_specialized_v on MSVC

It used to pretend that std::iterator_traits<T*> is a user-defined
specialization, which isn't the case. This is due to MSVC's
iterator_traits<T*> specialization not posing as the base template.