Remove ABI-incompatible assertion code in any_view.hpp (#1816)
* Remove ABI-incompatible assertion code in any_view.hpp

If a provider of an any_view is built without NDEBUG defined, but a
consumer is built with NDEBUG defined, the consumer will crash due to
ABI-incompatibility in any_ref.

Repros on clang with sanitizers, and that's an easily shareable demo:
  https://godbolt.org/z/Go9qxnjrM

Can't get it to repro with a small example in gcc, even with sanitizers

Repros in a small example on windows with msvc using same cpp/hpp files
as above link - usually assertion failure in the assertion that this
commit removes, sometimes Access Violation, but always crashes one way
or another. No sanitizers needed to get a crash with msvc. It's just
not easily shareable.

---------

Co-authored-by: Matt Beardsley <mbeardsl@mathworks.com>
Co-authored-by: Eric Niebler <eniebler@nvidia.com>